services/event_manager/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
app==0.0.1
bcrypt==4.0.1
beanie==1.29.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lazy-model==0.2.0
limits==4.6
motor==3.7.0
multidict==6.3.2
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.11.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.1
testcontainers==4.10.0
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
